requirements.txt

Markdown==2.6.2
MySQL-python==1.2.3
python-mysqldb
#---------------------------------------------------------------------------------

# Base
Django==1.8
celery==3.1.18
django-celery==3.1.16
django-braces==1.8.1
django-widget-tweaks==1.3
djangorestframework==3.1.3
psutil==3.0.1
python-dateutil==2.4.2
pytz==2015.4
requests==2.2.1
wolframalpha==1.2
django-secure==1.0.1
#django-debug-toolbar
django-angular

#---------------------------------------------------------------------------------

# development
# apt-get install graphviz graphviz-dev pkg-config ##debian graph of models
# pip install pygraphviz # debian graph of models
django-extensions==1.5.5

#---------------------------------------------------------------------------------
# Heroku production
#whitenoise
#gunicorn==19.3.0
#django-postgrespool
#dj-database-url==0.2.2